Refactored and updated test suite structure and loading strategy to allow for more modular test organization.
