+++ test/suite/controllers/types/Controller/test.js
@@ -1,0 +1,9 @@
+const classConstructor = require("./suite/constructor");
+
+describe("Controller", function() {
+
+    const addDesc = ( name, suite ) => describe(name, suite.bind(this));
+
+    addDesc( "constructor()", classConstructor );
+
+});
